+++ shortcut_composer/templates/pie_menu_utils/label_widget_utils/__init__.py
@@ -1,0 +1,3 @@
+from .create_label_widget import create_label_widget
+
+__all__ = ["create_label_widget"]
