+++ shortcut_composer/templates/pie_menu_utils/label_widget_utils/icon_label_widget.py
