+++ shortcut_composer/templates/pie_menu_utils/pie_manager.py
@@ -8,9 +8,8 @@ from PyQt5.QtGui import QCursor
 from api_krita.pyqt import Timer
 from composer_utils import Config
 from .pie_widget import PieWidget
-from .label_widgets import LabelWidget
-from .circle_points import CirclePoints
-from .label_animator import LabelAnimator
+from .label_widget import LabelWidget
+from .widget_utils import CirclePoints
 
 
 class PieManager:
@@ -61,3 +60,34 @@ class PieManager:
         if self._holder.active != widget:
             self._holder.active = widget
             self._animator.start()
+
+
+class LabelAnimator:
+    """
+    Controls the animation of background under pie labels.
+
+    Handles the whole widget at once, to prevent unnecessary repaints.
+    """
+
+    def __init__(self, widget: PieWidget) -> None:
+        self._widget = widget
+        self._children = widget.widget_holder
+        self._timer = Timer(self._update, Config.get_sleep_time())
+
+    def start(self):
+        """Start animating. The animation will stop automatically."""
+        self._timer.start()
+
+    def _update(self):
+        """Move all labels to next animation state. End animation if needed."""
+        for widget in self._children:
+            if self._children.active == widget:
+                widget.label.activation_progress.up()
+            else:
+                widget.label.activation_progress.down()
+
+        self._widget.repaint()
+        for widget in self._children:
+            if widget.label.activation_progress.value not in (0, 1):
+                return
+        self._timer.stop()
