+++ shortcut_composer/templates/pie_menu_utils/widget_utils/__init__.py
@@ -1,0 +1,9 @@
+from .circle_points import CirclePoints
+from .widget_holder import WidgetHolder
+from .pie_painter import PiePainter
+
+__all__ = [
+    "CirclePoints",
+    "WidgetHolder",
+    "PiePainter",
+]
